Deactivating social media section
Due to node.js version 16 being deprecated we can no longer risk triggering new deploys. 

New deploys are needed to display the instagram feed. Why that is:

Displaying instagram images requires an instagram token to be refreshed every 60 days. I've been handling this in postman but if that deadline is missed the key turns invalid and one needs to generate a new key, add it to environment variables, and trigger a new deploy. 

As node.js version 16 has been deprecated since 2024 this feels more and more risky, and we've therefore chosen to remove this element and freeze further deploys in order to keep the site build intact.

When running the project locally and trying to audit and update node.js to current version there are many errors which hopefully we can take a look at in the future.